Moved Effect to model
