+++ SuperCardGame/app/src/main/java/cs446_w2018_group3/supercardgame/model/buffs/DamageOverTimeBuff.java
@@ -1,6 +1,6 @@
 package cs446_w2018_group3.supercardgame.model.buffs;
 
-import cs446_w2018_group3.supercardgame.gamelogic.Effect;
+import cs446_w2018_group3.supercardgame.model.Effect;
 import cs446_w2018_group3.supercardgame.model.Player;
 
 /**
